refactor(config)!: Rename the `cli` package to `config`
BREAKING CHANGE: Modules previously imported from `powerapi.cli` must
now be imported from `powerapi.config`.
