+++ src/powerapi/config/binding_manager.py
