+++ src/powerapi/config/generator.py
